style(Header): update Header style
