+++ src/components/Header.module.css
@@ -1,71 +1,30 @@
+@import "tailwindcss";
+@reference "@/index.css";
+
 .header {
-  width: 100vw;
-  height: 150px;
-  background-color: lightgreen;
-  position: fixed;
-  z-index: 1000;
+  @apply flex w-full bg-lightGreen px-6 py-3 justify-center sticky top-0 z-5;
 }
 
-.header_top {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 100px;
+.headerContent {
+  @apply max-w-7xl justify-between flex grow;
 }
 
-.page_logo {
-  height: 50px;
-  width: auto;
-  margin-left:30px;
-}
-
-.header_title {
-  font-size: 3em;
-  color: #333;
-  white-space: nowrap;
-  overflow:hidden;
-  text-overflow: ellipsis;
-  text-align: center;
-}
-
-.user_icon {
-  height: 40px;
-  margin-right: 50px;
-  cursor: pointer;
-}
-
-.header_bottom {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0px 20px 0px;
-}
-
-.sideBar_icon {
-  height: 40px;
-  cursor: pointer;
-  margin-bottom: 10px;
+.headerContainer {
+  @apply flex gap-4 items-center;
 }
 
 .username {
-  width: fit-content;
-  font-size: 1.5em;
-  color: #333;
-  white-space:nowrap;
+  @apply h-full w-fit size-1.5 text-2xl text-gray-800 whitespace-nowrap;
 }
 
-.sideBarHoverArea {
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 50px;
-  z-index: 999;
+.hello {
+  @apply font-bold;
 }
 
-@media (max-width: 768px) {
-  .header {
-    height: auto;
-    flex-direction:row;
-  }
+.usernameIconContainer {
+  @apply relative;
+}
+
+.usernameIcon {
+  @apply cursor-pointer;
 }
